% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisionFunction.R
\name{add.alpha}
\alias{add.alpha}
\title{Add transparency to a color}
\usage{
add.alpha(col, alpha = 1)
}
\arguments{
\item{col}{A color.}

\item{alpha}{The level of transparency.}
}
\value{
The transparent color.
}
\description{
Add transparency to a color
}
\examples{
color1 <- add.alpha(col = "red", alpha = 0.6)

color2 <- add.alpha(colors()[115], 0.5)
x <- rnorm(1000, mean = 1, sd = 0.5)
y <- rnorm(1000, mean = 1, sd = 0.5)
plot(x,y, col = c(color1, color2), pch = 16)

}
\seealso{
\code{\link{rgb}}, \code{\link{colors}}
}
