% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Clones_E12_P0}
\alias{Clones_E12_P0}
\title{Data from observed clones and their properties after electroporation at E12 and sacrifice at P0}
\format{A data frame with 63 rows (clones) and 14 variables:
\describe{
  \item{animal_n}{Identifier of the animal the clone was observed in}
  \item{clone_ID}{Identifier of the clone, combination of the different colors observed after recombination}
  \item{time_of_electroporation_}{Development time when the electroporation was performed}
  \item{time_of_analysis}{Time when the sacrifice and analysis were performed}
  \item{clone_type}{Type of the clone, either homogeneous (HomC = clone contains cells of the same type), or heterogeneous (HetC = clone contains cells of more than one type)}
  \item{clone_subtype}{Subtype of the clone either Cortical (all astrocytes are located in the cortex), WM (all astrocytes in the white matter), Cortical_WMA (astrocytes both in the cortex and in the white matter)}
  \item{clone_subtype.1}{Same definition as clone_subtype but here the detail of the cortical astrocyte is given (BG progenitor or GLA)}
  \item{M__L_dispersion_.µm.}{Mediolateral dispersion of the clone, in µm}
  \item{clone_size}{Number of cells belonging to the clone (cells expressing the same color combination)}
  \item{number_of_BG}{Number of Bergmann Glia cells}
  \item{number_of_GLA}{Number of Granular Cell layer Astrocytes}
  \item{number_of_WMA}{Number of White Matter Astrocytes}
  \item{number_of_VZA}{Number of Ventricular Zone Astrocyte}
  \item{folium}{The folium where the clone is located}
}}
\source{
Cerrato et al 2018
}
\usage{
Clones_E12_P0
}
\description{
Data from observed clones and their properties after electroporation at E12 and sacrifice at P0
}
\keyword{datasets}
