% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Clones_E12_P30}
\alias{Clones_E12_P30}
\title{Data from observed clones and their properties after electroporation at E12 and sacrifice at P30}
\format{A data frame with 285 rows (clones) and 14 variables:
\describe{
  \item{Animal_n}{Identifier of the animal the clone was observed in}
  \item{clone_ID}{Identifier of the clone, combination of the different colors observed after recombination}
  \item{time_of_electroporation_}{Development time when the electroporation was performed}
  \item{time_of_analysis}{Time when the sacrifice and analysis were performed}
  \item{clone_type}{Type of the clone, either homogeneous (HomC = clone contains cells of the same type), or heterogeneous (HetC = clone contains cells of more than one type)}
  \item{clone_subtype}{Subtype of the clone based on the astrocyte types present}
  \item{M__L_dispersion_.µm.}{Mediolateral dispersion of the clone, in µm}
  \item{clone_size}{Number of cells belonging to the clone (cells expressing the same color combination)}
  \item{number_of_BG}{Number of Bergmann Glia cells}
  \item{number_of_GLA}{Number of Granular Cell layer Astrocytes}
  \item{number_of_WMA}{Number of White Matter Astrocytes}
  \item{number_of_CNA}{Number of Cerebellar Nuclei Astrocytes}
  \item{number_of_VZA}{Number of Ventricular Zone Astrocyte}
  \item{localization}{Comments about the localization of the clone in the cerebellum}
}}
\source{
Cerrato et al 2018
}
\usage{
Clones_E12_P30
}
\description{
Data from observed clones and their properties after electroporation at E12 and sacrifice at P30
}
\keyword{datasets}
