% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cloneTable}
\alias{cloneTable}
\title{Example of initialization of the cloneTable for the divisionMPGenerationDpd function}
\format{A dataframe with 4 columns containing initially the data for the 1st mother cell:
\describe{
  \item{motherID}{ID of the mother of the current cell}
  \item{cellID}{ID of the current cell}
  \item{type}{Type of the current cell}
  \item{timepoint}{Generation the current cell is born. Mother cell is initialized at generation 1}
}}
\source{
Cerrato et al 2018
}
\usage{
cloneTable
}
\description{
Example of initialization of the cloneTable for the divisionMPGenerationDpd function
}
\keyword{datasets}
