% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{currentCell}
\alias{currentCell}
\title{Example of initialization of the first mother cell for the divisionMPGenerationDpd function}
\format{A dataframe with 1row and 4 columns containing the data for the 1st mother cell:
\describe{
  \item{cellID}{ID of the 1st cell of the lineage}
  \item{type}{Type of the 1st cell of the lineage}
  \item{timepoint}{Set to generation 1}
}}
\source{
Cerrato et al 2018
}
\usage{
currentCell
}
\description{
Example of initialization of the first mother cell for the divisionMPGenerationDpd function
}
\keyword{datasets}
