% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisionFunction.R
\name{divisionMPGenerationDpd}
\alias{divisionMPGenerationDpd}
\title{Simulate an astrocyte lineage}
\usage{
divisionMPGenerationDpd(motherCell = currentCell, cloneOutput = cloneTable,
  transitionMatrix = list(transition.Progenitor, transition.Astro,
  firstPostMitoticCell), maxCount = 1e+05, parameterInterpolation = T)
}
\arguments{
\item{motherCell}{the information table of the mother cell}

\item{cloneOutput}{the initialized clone table with the 1st cell information}

\item{transitionMatrix}{a list specifying the transition matrix of the particular progenitor type:

 - transition.Progenitor : probabilities of cell cycle exit vs amplification of the progenitor

 - transition.Astro : probabilities of generating the different astrocyte types after cell cycle exit

 - firstPostMitoticCell : for the one-cell clones, proportions of the observed astrocyte types}

\item{maxCount}{number of trials after which the function will stop if the lineage never stop (stochasticity issue)}

\item{parameterInterpolation}{boolean indicating whether interpolation was performed on the input data (always TRUE)}
}
\description{
Simulate an astrocyte lineage
}
\examples{


}
