% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisionFunction.R
\name{error.bar}
\alias{error.bar}
\title{Add error bars to a bar plot}
\usage{
error.bar(x, y, upper, lower = upper, length = 0.05, ...)
}
\arguments{
\item{x}{the barplot x coordinates.}

\item{y}{the data (height of the bars).}

\item{upper}{matrix of the different error bar values.}

\item{lower}{matrix of the different error bar values. By default equals to upper}

\item{length}{length of the horizontal lines of the error bars}
}
\description{
Add error bars to a bar plot
}
\examples{
v1  <- c(1:4)
my.x <- barplot(v1)
error <- c(0.2, 1, 0.5, 0.1)
error.bar(x = my.x, y = v1, upper = error)

}
