% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{firstPostMitoticCell}
\alias{firstPostMitoticCell}
\title{Example of parameters used to govern the fate of postmitotic astrocyte directly targeted by the electroporation (1 cell clone)}
\format{A list of 2 elements:
\describe{
  \item{probability}{Vector: probabilities of the one cell clone to be a BG (1st column), a GLA (2nd column), or a WMA (3rd column)}
  \item{type}{The corresponding names of different possible outcome, in the same order as the columns of the probability matrix (BG, GLA, WMA) 
              (number of types correponds to order of the probability vector element)}
}}
\source{
Cerrato et al 2018
}
\usage{
firstPostMitoticCell
}
\description{
Example of parameters used to govern the fate of postmitotic astrocyte directly targeted by the electroporation (1 cell clone)
}
\keyword{datasets}
