% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transition.MP}
\alias{transition.MP}
\title{Example of parameters used to govern multipotent progenitors behavior}
\format{A list of 3 elements:
\describe{
  \item{probability}{Matrix with columns: probabilities of generating a MP (1st column), or an Astrocyte (2nd column), 
  
                                  rows :  for each generation in the simulated lineage (number of rows can vary)}
  \item{type}{The corresponding names of different possible outcome, in the same order as the columns of the probability matrix (MP, Astro) 
              (number of types correponds to the number of columns of the above probability matrix)}
  \item{generationInterval}{Vector of length corresponding to the number of generations, length equal to the number of rows of the matrix probability}
}}
\source{
Cerrato et al 2018
}
\usage{
transition.MP
}
\description{
Example of parameters used to govern multipotent progenitors behavior
}
\keyword{datasets}
